Added Ball Scalability
Ball now scales whenever it is picked up and thrown. The Scale factor is a public variable that can be tweaked.

Currently any object picked up will scale; could add tags and checks to make it such that only the ball will scale but that may slow the script and delay the collision toggle.